Reorganized file structure (move sections to components directory
